Ignore matplotlib axes when caching
These do not impact computations, and are used for visualization only.

Note that this is a stop-gap measure to make the pipeline run.  It
will break the generation of some plots.  If you wish to re-run
plots, remove the `cachedir` first.

Ultimately, a solution is to split visualization and computation routines.